I readded the separate minimum player messages (I accidentally deleted them before)
the qt themes in the main file were being overridden by the ones in the utils file so I moved that logic to the main file. I removed leftover code from the old legacy interface. the main file theme logic was simplified as windowsvista is the only theme that works properly with the stylesheet.  Themes like fusion and windows will break the interface is there dark mode themes changing the color of some icons to create contrast with the dark interface but since the stylesheet is light mode the icons disappear in the white background